add a function that plots a pole with a dp_dm ellipse
Determining the angle to rotate requires applying the two laws of cosines for spherical triangles so that dp is aligned with the great circle between the pole and the site. A number of conditional statements are implemented to determine what direction to apply the associated angle. Hopefully these are robust to all cases.