change form_Mhat to take mhat not X
The version of form_Mhat() committed to pmag.py took
X as input (collection of directions). However, the
Heslop et al. 2023 code is set up for the input to this function
to be the mhat vector. This commit changes the input to as in the
provided Heslop et al. 2023 code.