Add mean_bootstrap_confidence() and plot_bootstrap_confidence() functions
This commit adds two new functions to ipmag.py: mean_bootstrap_confidence() and plot_bootstrap_confidence(). These functions are used to calculate and plot bootstrap confidence intervals for the mean direction of a set of directions. The functions are based on the bootstrap functions in the pmagpy package, but have been modified to work with the new directional data class.
The functions using the pmag.form_Mhat form_Ghat