add common_mean_bootstrap_H23()
This commit adds a common_mean_bootstrap_H23() function
to ipmag.py. The code for this function comes from Heslop et al.
2023 (https://doi.org/10.5281/zenodo.8141070) with some minor
bells and whistles related to the plotting and return statements
in addition to getting it all within a function. There is also a
reversal test wrapper function as is the structure for
common_mean_bootstrap() and reversal_test_bootstrap(). This function
follows the dec, inc, di_block structure of other ipmag functions.
That function is reversal_test_bootstrap_H23().